requirements.txt
Flask
pymongo
python-dotenv
requests
bcrypt
PyJWT
gridfs 